refactoring InnerState class
