+++ lib/helpers/column-set.js
@@ -7,7 +7,7 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-const {InnerState} = require('../utils/inner-state');
+const {InnerState} = require('../inner-state');
 const {assertOptions} = require('assert-options');
 const {TableName} = require('./table-name');
 const {Column} = require('./column');
