+++ lib/inner-state.js
@@ -1,25 +1,25 @@
-const {addReadProp} = require('./');
+const {addReadProp} = require('./utils');
 
 /**
  * @private
  * @class InnerState
  * @description
- * Implements generic implementation for supporting private/inner state inside classes.
- *
- * A derived class can access its inner state via hidden read-only property _inner.
+ * Implements support for private/inner state object inside the class,
+ * which can be accessed by a derived class via hidden read-only property _inner.
  */
 class InnerState {
 
     constructor(initialState) {
-        let state = {};
+        addReadProp(this, '_inner', {}, true);
         if (initialState && typeof initialState === 'object') {
-            state = Object.assign({}, initialState);
+            this.extendState(initialState);
         }
-        addReadProp(this, '_inner', state, true);
     }
 
     /**
      * Extends or overrides inner state with the specified properties.
+     *
+     * Only own properties are used, i.e. inherited ones are skipped.
      */
     extendState(state) {
         for (const a in state) {
