+++ lib/query-file.js
@@ -7,7 +7,7 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-const {InnerState} = require('./utils/inner-state');
+const {InnerState} = require('./inner-state');
 const {QueryFileError} = require('./errors');
 const {assertOptions} = require('assert-options');
