+++ lib/tx-mode.js
@@ -7,9 +7,9 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-const {InnerState} = require('./utils/inner-state');
-const {assertOptions} = require('assert-options');
+const {InnerState} = require('./inner-state');
 const {addInspection} = require('./utils');
+const {assertOptions} = require('assert-options');
 
 /**
  * @enum {number}
