+++ lib/types/server-formatting.js
@@ -1,4 +1,4 @@
-const {InnerState} = require('../utils/inner-state');
+const {InnerState} = require('../inner-state');
 const {addInspection} = require('../utils');
 const utils = require('../utils');
